refactor(cdk): fix file name
